fix: Change user_id column type from INT to BIGINT
NextAuth generates large user IDs that exceed INT range.
Changed all user_id columns to BIGINT to support these IDs.

Changes:
- Updated create_push_notifications.sql schema
- Added migration script for existing tables
- Updated setup documentation with migration instructions

+++ sql/create_push_notifications.sql
@@ -1,7 +1,7 @@
 -- プッシュ通知サブスクリプションテーブル
 CREATE TABLE IF NOT EXISTS push_subscriptions (
   id INT AUTO_INCREMENT PRIMARY KEY,
-  user_id INT NOT NULL,
+  user_id BIGINT NOT NULL,
   endpoint TEXT NOT NULL,
   p256dh VARCHAR(255) NOT NULL,
   auth VARCHAR(255) NOT NULL,
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS push_subscriptions (
 -- 通知設定テーブル
 CREATE TABLE IF NOT EXISTS notification_settings (
   id INT AUTO_INCREMENT PRIMARY KEY,
-  user_id INT NOT NULL UNIQUE,
+  user_id BIGINT NOT NULL UNIQUE,
   notifications_enabled BOOLEAN DEFAULT TRUE,
   advance_before_start BOOLEAN DEFAULT TRUE,
   raffle_on_start BOOLEAN DEFAULT TRUE,
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS notification_settings (
 -- 通知履歴テーブル（重複防止用）
 CREATE TABLE IF NOT EXISTS notification_history (
   id INT AUTO_INCREMENT PRIMARY KEY,
-  user_id INT NOT NULL,
+  user_id BIGINT NOT NULL,
   event_id VARCHAR(255) NOT NULL,
   notification_type ENUM('advance_start', 'raffle_start', 'raffle_end') NOT NULL,
   sent_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
